change from pht
sound, widget error, random card position, delete ad code relate, remove useless code
