+++ pubspec.yaml
@@ -67,9 +67,10 @@ flutter:
   # the material Icons class.
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
+  # To add assets to your application, add an asset section, like this:
   assets:
-    - asset/images/
+    - assets/images/
+    - assets/audios/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
@@ -86,7 +87,7 @@ flutter:
   fonts:
     - family: Source
       fonts:
-        - asset: asset/fonts/SourceSansPro-Regular.ttf
+        - asset: assets/fonts/SourceSansPro-Regular.ttf
   #       - asset: fonts/Schyler-Italic.ttf
   #         style: italic
   #   - family: Trajan Pro
